#0000 Fixed size bug in fdf_multi_test TESTS_RAN: fdf_multi_test
git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2491 01a69087-22d5-4a29-8152-8a1d5e10e5e9

+++ api/tests/fdf_multi_test.c
@@ -43,7 +43,7 @@ int get_options(int argc, char *argv[])
         switch (c) {
 
         case 's':
-            size = atoi(optarg) * 1024 * 1024 * 1024;
+            size = atol(optarg) * 1024 * 1024 * 1024;
             break;
 
         case 'i':
